sp: addtional ==> additional
