+++ src/io/flutter/module/FlutterSmallIDEGeneratorPeer.java
@@ -15,7 +15,7 @@ import com.intellij.platform.WebProjectGenerator;
 import com.intellij.ui.ComboboxWithBrowseButton;
 import com.intellij.ui.DocumentAdapter;
 import io.flutter.FlutterBundle;
-import io.flutter.module.settings.FlutterCreateAddtionalSettingsFields;
+import io.flutter.module.settings.FlutterCreateAdditionalSettingsFields;
 import io.flutter.sdk.FlutterCreateAdditionalSettings;
 import io.flutter.sdk.FlutterSdkUtil;
 import org.jetbrains.annotations.NotNull;
@@ -30,7 +30,7 @@ import javax.swing.text.JTextComponent;
 
 public class FlutterSmallIDEGeneratorPeer implements WebProjectGenerator.GeneratorPeer<String> {
   private final ComboboxWithBrowseButton sdkPathComboWithBrowse;
-  private final FlutterCreateAddtionalSettingsFields settingsFields;
+  private final FlutterCreateAdditionalSettingsFields settingsFields;
 
   public FlutterSmallIDEGeneratorPeer() {
     sdkPathComboWithBrowse = new ComboboxWithBrowseButton(new ComboBox<>());
@@ -42,7 +42,7 @@ public class FlutterSmallIDEGeneratorPeer implements WebProjectGenerator.Generat
       FileChooserDescriptorFactory.createSingleFolderDescriptor(),
       TextComponentAccessor.STRING_COMBOBOX_WHOLE_TEXT);
 
-    settingsFields = new FlutterCreateAddtionalSettingsFields();
+    settingsFields = new FlutterCreateAdditionalSettingsFields();
   }
 
   @NotNull
@@ -100,8 +100,8 @@ public class FlutterSmallIDEGeneratorPeer implements WebProjectGenerator.Generat
   }
 
   @NotNull
-  public FlutterCreateAdditionalSettings getAddtionalSettings() {
-    return settingsFields.getAddtionalSettings();
+  public FlutterCreateAdditionalSettings getAdditionalSettings() {
+    return settingsFields.getAdditionalSettings();
   }
 
   @NotNull
